Move win32/user to win32/net
