+++ lib/chef/win32/net.rb
